+++ compose/mpp/demo/src/commonMain/kotlin/androidx/compose/mpp/demo/App.kt
@@ -33,7 +33,6 @@ val MainScreen = Screen.Selection(
     LazyLayouts,
     TextFields,
     AndroidTextFieldSamples,
-    Screen.Example("IosKeyboardBugExample") { IosKeyboardBugExample() },
 )
 
 sealed interface Screen {
@@ -51,7 +50,7 @@ sealed interface Screen {
         constructor(title: String, vararg screens: Screen) : this(title, listOf(*screens))
 
         fun mergedWith(screens: List<Screen>): Selection {
-            return Selection(title, this.screens + screens)
+            return Selection(title, screens + this.screens)
         }
     }
