+++ compose/mpp/demo/src/uikitMain/kotlin/bugs/KeyboardEmptyWhiteSpace.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package androidx.compose.mpp.demo
+package bugs
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
@@ -32,7 +32,7 @@ import androidx.compose.material.Button
 import androidx.compose.material.OutlinedTextField
 import androidx.compose.material.Text
 import androidx.compose.material.TextFieldDefaults
-import androidx.compose.runtime.Composable
+import androidx.compose.mpp.demo.Screen
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -41,9 +41,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 
-@Composable
-fun IosKeyboardBugExample() {
-    // Reproducer of bugs, described here: https://github.com/JetBrains/compose-multiplatform/issues/3490
+val KeyboardEmptyWhiteSpace = Screen.Example("KeyboardEmptyWhiteSpace") {
+    // Issue: https://github.com/JetBrains/compose-multiplatform/issues/3490
     var messages by remember { mutableStateOf(listOf<String>()) }
     Column(
         modifier = Modifier.fillMaxSize().background(Color.Black)
@@ -82,4 +81,5 @@ fun IosKeyboardBugExample() {
             }
         }
     }
+
 }
